Moved AppState and all versions to separate file
