+++ src/reducers/AppState.ts
@@ -1,0 +1,5 @@
+import { AppStateV2 } from './version2';
+
+export const currentAppStateVersion = 2;
+
+export type AppState = AppStateV2;
